course-generator/requirements.txt
# requirements.txt - Dependencies
fastapi
uvicorn
openai
python-dotenv